// upvertice-ui/src/app/services/models/page-response-sponsor-offer-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { SponsorOfferResponse } from '../models/sponsor-offer-response';
export interface PageResponseSponsorOfferResponse {
  content?: Array<SponsorOfferResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}