// upvertice-ui/src/app/services/models/page-response-sponsorship-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { SponsorshipResponse } from '../models/sponsorship-response';
export interface PageResponseSponsorshipResponse {
  content?: Array<SponsorshipResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}