// upvertice-ui/src/app/services/models/sponsor-offer-multipart-choose-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface SponsorOfferMultipartChooseRequest {

  /**
   * Optional image files for the ad design
   */
  images?: Array<Blob>;

  /**
   * SponsorAdRequest serialized as JSON string
   */
  request?: string;
}